Integrate archival api (#2)
* add archival back end

* add archival status

* add full path for migrate

* Do not remove file if backend if archival

* implement readat for archival backend

* fix log string error

* use async api and add process data

* support dm async apis

+++ weed/server/volume_grpc_tier_upload.go
@@ -8,6 +8,7 @@ import (
 	"github.com/chrislusf/seaweedfs/weed/pb/volume_server_pb"
 	"github.com/chrislusf/seaweedfs/weed/storage/backend"
 	"github.com/chrislusf/seaweedfs/weed/storage/needle"
+	"github.com/chrislusf/seaweedfs/weed/glog"
 )
 
 // VolumeTierMoveDatToRemote copy dat file to a remote tier
@@ -63,7 +64,9 @@ func (vs *VolumeServer) VolumeTierMoveDatToRemote(req *volume_server_pb.VolumeTi
 	}
 
 	// copy the data file
-	key, size, err := backendStorage.CopyFile(diskFile.File, fn)
+	fullName := v.FileName(".dat")
+	glog.V(0).Infof("start copying file: %s", fullName, " to backend type:%s", backendType)
+	key, size, err := backendStorage.CopyFile(fullName, diskFile.File, fn)
 	if err != nil {
 		return fmt.Errorf("backend %s copy file %s: %v", req.DestinationBackendName, diskFile.Name(), err)
 	}
@@ -87,7 +90,7 @@ func (vs *VolumeServer) VolumeTierMoveDatToRemote(req *volume_server_pb.VolumeTi
 		return fmt.Errorf("volume %d fail to load remote file: %v", v.Id, err)
 	}
 
-	if !req.KeepLocalDatFile {
+	if !req.KeepLocalDatFile && backendType != "archival" {
 		os.Remove(v.FileName(".dat"))
 	}
